# refactor: do not allocate for SanitizedCopy()