**What this PR does / why we need it**:

`SanitizedCopy()` functions unnecessarily allocate objects where it's not needed.

This PR removes those allocations in several places and adds a benchmark to prove the results of this optimization:

before

```
go test -benchmem -bench BenchmarkSanitizedCopy -count 1 ./internal/dataplane/kongstate
goos: darwin
goarch: arm64
pkg: github.com/kong/kubernetes-ingress-controller/v3/internal/dataplane/kongstate
cpu: Apple M1 Max
BenchmarkSanitizedCopy-10           1837            609341 ns/op         3216834 B/op       1036 allocs/op
PASS
ok      github.com/kong/kubernetes-ingress-controller/v3/internal/dataplane/kongstate   1.788s
```

after

```
go test -benchmem -bench BenchmarkSanitizedCopy -count 1 ./internal/dataplane/kongstate
goos: darwin
goarch: arm64
pkg: github.com/kong/kubernetes-ingress-controller/v3/internal/dataplane/kongstate
cpu: Apple M1 Max
BenchmarkSanitizedCopy-10           4982            242434 ns/op          778484 B/op          5 allocs/op
PASS
ok      github.com/kong/kubernetes-ingress-controller/v3/internal/dataplane/kongstate   2.750s
```
 
```
benchstat -confidence 0.9 old new
goos: darwin
goarch: arm64
pkg: github.com/kong/kubernetes-ingress-controller/v3/internal/dataplane/kongstate
cpu: Apple M1 Max
                 │     old     │                new                 │
                 │   sec/op    │   sec/op     vs base               │
SanitizedCopy-10   624.5µ ± 1%   241.3µ ± 1%  -61.37% (p=0.008 n=5)

                 │      old      │                 new                 │
                 │     B/op      │     B/op      vs base               │
SanitizedCopy-10   3141.4Ki ± 0%   760.2Ki ± 0%  -75.80% (p=0.008 n=5)

                 │      old      │                new                │
                 │   allocs/op   │ allocs/op   vs base               │
SanitizedCopy-10   1036.000 ± 0%   5.000 ± 0%  -99.52% (p=0.008 n=5)
```

## Special notes 


This can be taken a step further by removing the allocations for credentials i.e. https://github.com/Kong/kubernetes-ingress-controller/blob/4b60bf8cd8074fb310b8270e623fe8f080cdd934/internal/dataplane/kongstate/credentials.go#L163-L234 but that requires a bit more refactoring.